Narrow FeatherPGS to fused Warp
Remove alternate PGS modes and dead kernels so the private solver has a single matrix-free Warp path. Keep velocity limits always active for finite limits, add body_parent_f support, and cover unsupported kinematic bodies.
